+++ level3.cpp
@@ -11,7 +11,7 @@ pair<pair<int, int>, pair<int, int>> Level3::generateMove(string input) {
         for (int j = 0; j < 8; j++) {
             if ((*boardmap)[i][j] && (*boardmap)[i][j]->getColor() == *currentPlayer) {
                 vector<pair<int, int>> legalmove = (*boardmap)[i][j]->legalMoves(i, j);
-                for (int k = 0; k < legalmove.size(); k++) {
+                for (size_t k = 0; k < legalmove.size(); k++) {
                     pair<pair<int, int>, pair<int, int>> temp = {{i,                  j},
                                                                  {legalmove[k].first, legalmove[k].second}};
                     legalmoves.emplace_back(temp);
@@ -25,7 +25,7 @@ pair<pair<int, int>, pair<int, int>> Level3::generateMove(string input) {
         for (int j = 0; j < 8; j++) {
             if ((*boardmap)[i][j] && (*boardmap)[i][j]->getColor() == *currentPlayer) {
                 vector<pair<int, int>> capmove = (*boardmap)[i][j]->captureMoves(i, j);
-                for (int k = 0; k < capmove.size(); k++) {
+                for (size_t k = 0; k < capmove.size(); k++) {
                     pair<pair<int, int>, pair<int, int>> temp = {{i,                j},
                                                                  {capmove[k].first, capmove[k].second}};
                     capmoves.emplace_back(temp);
@@ -39,7 +39,7 @@ pair<pair<int, int>, pair<int, int>> Level3::generateMove(string input) {
         for (int j = 0; j < 8; j++) {
             if ((*boardmap)[i][j] && (*boardmap)[i][j]->getColor() == *currentPlayer) {
                 vector<pair<int, int>> checkmove = (*boardmap)[i][j]->checkMoves(i, j);
-                for (int k = 0; k < checkmove.size(); k++) {
+                for (size_t k = 0; k < checkmove.size(); k++) {
                     pair<pair<int, int>, pair<int, int>> temp = {{i,                  j},
                                                                  {checkmove[k].first, checkmove[k].second}};
                     checkmoves.emplace_back(temp);
@@ -53,7 +53,7 @@ pair<pair<int, int>, pair<int, int>> Level3::generateMove(string input) {
         for (int j = 0; j < 8; j++) {
             if ((*boardmap)[i][j] && (*boardmap)[i][j]->getColor() == *currentPlayer) {
                 vector<pair<int, int>> avoidmove = (*boardmap)[i][j]->avoidMoves(i, j);
-                for (int k = 0; k < avoidmove.size(); k++) {
+                for (size_t k = 0; k < avoidmove.size(); k++) {
                     pair<pair<int, int>, pair<int, int>> temp = {{i,                  j},
                                                                  {avoidmove[k].first, avoidmove[k].second}};
                     avoidmoves.emplace_back(temp);
